chore(backup-viewer): Clean up code
